V0.9.4 (#24)
* ensure --dryrun supercedes run
* add alternate env inventory
* fix ssh-keygen once and for all
* fix --delete-disks usage

+++ bin/tdh-workers-init.sh
@@ -11,7 +11,7 @@ fi
 
 # -----------------------------------
 
-names="d01 d02 d03"
+names="d01 d02 d03 d04"
 prefix="$TDH_GCP_PREFIX"
 
 zone=
@@ -25,7 +25,7 @@ subnet="tdh-net-west1"
 
 myid=1
 attach=0
-dryrun=1
+dryrun=0
 ssd=0
 action=
 rt=
@@ -152,28 +152,28 @@ if [ -n "$network" ] && [ -z "$subnet" ]; then
     exit 1
 fi
 
+echo ""
+version
+
+if [ "$action" == "run" ] && [ $dryrun -eq 0 ]; then
+    dryrun=0
+else
+    dryrun=1  # action -ne run
+    echo "  <DRYRUN> enabled"
+fi
+
 if [ -n "$zone" ]; then
     gssh="$gssh --zone $zone"
     gscp="$gscp --zone $zone"
 fi
 
-echo "" 
-version
-echo ""
-
-if [ "$action" == "run" ]; then
-    dryrun=0
-else
-    echo "  <DRYRUN> enabled"
-fi
-
 if [ -n "$namelist" ]; then
     names="$namelist"
 else
-    echo "Using default 3 workers"
+    echo "Using default of 4 worker instances"
 fi
 
-echo "Creating a worker instance '$mtype' for { $names }"
+echo "Creating worker instance '$mtype' for { $names }"
 echo ""
 
 
@@ -181,10 +181,10 @@ for name in $names; do
     #
     # Create instance
     host="${prefix}-${name}"
-    cmd="${tdh_path}/tdh-gcp-compute.sh --prefix ${prefix} --type ${mtype} --bootsize ${bootsize}"
+    cmd="${tdh_path}/tdh-gcp-compute.sh --prefix $prefix --type $mtype --bootsize $bootsize"
     
     if [ -n "$network" ]; then
-        cmd="$cmd --network ${network} --subnet ${subnet}"
+        cmd="$cmd --network $network --subnet $subnet"
     fi
     if [ -n "$zone" ]; then
         cmd="$cmd --zone ${zone}"
@@ -193,7 +193,7 @@ for name in $names; do
         cmd="${cmd} --dryrun"
     fi
     if [ $attach -gt 0 ]; then
-        cmd="${cmd} --attach --disksize ${disksize}"
+        cmd="${cmd} --attach --disksize $disksize"
     fi
     if [ $ssd -gt 0 ]; then
         cmd="${cmd} --ssd"
@@ -211,16 +211,19 @@ for name in $names; do
     fi
 done
 
+
 if [ $rt -gt 0 ]; then
     exit $rt
 fi
+
+
 echo ""
 echo " -> Waiting for host to respond"
 
 if [ $dryrun -eq 0 ]; then
     sleep 10
     for x in {1..3}; do 
-        yf=$( $gssh ${host} --command 'uname -n' )
+        yf=$( $gssh $host --command 'uname -n' )
         if [[ $yf == $host ]]; then
             echo " It's ALIVE!!!"
             break
@@ -231,18 +234,22 @@ if [ $dryrun -eq 0 ]; then
 fi
 echo ""
 
+
 for name in $names; do
+    host="${prefix}-${name}"
+
     #
     # Device format and mount
-    host="${prefix}-${name}"
     if [ $attach -gt 0 ]; then
         device="/dev/sdb"
         mountpoint="/data1"
+
         echo "( $gssh ${host} --command './tdh-gcp-format.sh $device $mountpoint' )"
+
         if [ $dryrun -eq 0 ]; then
             ( $gscp ${tdh_path}/tdh-gcp-format.sh ${host}: )
-            ( $gssh ${host} --command 'chmod +x tdh-gcp-format.sh' )
-            ( $gssh ${host} --command "./tdh-gcp-format.sh $device $mountpoint" )
+            ( $gssh $host --command 'chmod +x tdh-gcp-format.sh' )
+            ( $gssh $host --command "./tdh-gcp-format.sh $device $mountpoint" )
         fi
 
         rt=$?
@@ -255,6 +262,7 @@ for name in $names; do
     #
     # disable  iptables and cups
     echo "( $gssh $host --command 'sudo systemctl stop firewalld; sudo systemctl disable firewalld; sudo service cups stop; sudo chkconfig cups off' )"
+
     if [ $dryrun -eq 0 ]; then
         ( $gssh $host --command "sudo systemctl stop firewalld; sudo systemctl disable firewalld; sudo service cups stop; sudo chkconfig cups off" )
     fi
@@ -262,12 +270,13 @@ for name in $names; do
 
     #
     # prereq's
-    echo "( $gssh ${host} --command ./tdh-prereqs.sh )"
+    echo "( $gssh $host --command ./tdh-prereqs.sh )"
+
     if [ $dryrun -eq 0 ]; then
         ( $gscp ${tdh_path}/../etc/bashrc ${host}:.bashrc )
         ( $gscp ${tdh_path}/tdh-prereqs.sh ${host}: )
-        ( $gssh ${host} --command 'chmod +x tdh-prereqs.sh' )
-        ( $gssh ${host} --command ./tdh-prereqs.sh )
+        ( $gssh $host --command 'chmod +x tdh-prereqs.sh' )
+        ( $gssh $host --command ./tdh-prereqs.sh )
     fi
 
     rt=$?
@@ -280,14 +289,15 @@ for name in $names; do
     #
     # ssh
     echo "( $gscp ${master_id_file} ${host}:.ssh" 
+
     if [ $dryrun -eq 0 ]; then
         ( $gscp ${master_id_file} ${host}:.ssh/ )
-        ( $gssh ${host} --command "cat .ssh/${master_id} >> .ssh/authorized_keys; chmod 700 .ssh; chmod 600 .ssh/authorized_keys " )
+        ( $gssh $host --command "cat .ssh/${master_id} >> .ssh/authorized_keys; chmod 700 .ssh; chmod 600 .ssh/authorized_keys " )
     fi
 
     # mysql client
     role="client"
-    cmd="$tdh_path/tdh-mysql-install.sh"
+    cmd="${tdh_path}/tdh-mysql-install.sh"
 
     if [ -n "$zone" ]; then
         cmd="$cmd --zone $zone"
@@ -295,7 +305,7 @@ for name in $names; do
 
     echo "( $cmd $host $role )"
     if [ $dryrun -eq 0 ]; then
-        ( ${cmd} ${host} ${role} )
+        ( $cmd $host $role )
     fi
     
     rt=$?
